refactor: page structure
