+++ components/Layout/index.tsx
@@ -1,0 +1,21 @@
+import Head from 'next/head';
+import { ReactNode } from 'react';
+
+interface LayoutProps {
+  children: ReactNode;
+  title: string;
+}
+
+export default function Layout({ children, title }: LayoutProps) {
+  return (
+    <>
+      <Head>
+        <title>{title}</title>
+        <meta property="og:title" content={title} />
+        <meta property="twitter:title" content={title} />
+      </Head>
+
+      <div className="content">{children}</div>
+    </>
+  );
+}
